Support Swift 5 and RxSwift 5.0' - Deprecated StateRelay. Use BehaviorRelay
