+++ Package.swift
@@ -1,4 +1,4 @@
-// swift-tools-version:4.2
+// swift-tools-version:5.0
 
 import PackageDescription
 
@@ -8,11 +8,14 @@ let package = Package(
     .library(name: "ReactorKit", targets: ["ReactorKit"]),
   ],
   dependencies: [
-    .package(url: "https://github.com/ReactiveX/RxSwift.git", .upToNextMajor(from: "4.0.0")),
-    .package(url: "https://github.com/devxoul/RxExpect.git", .upToNextMajor(from: "1.0.0"))
+    .package(url: "https://github.com/ReactiveX/RxSwift.git", .upToNextMajor(from: "5.0.0")),
+
+    // TODO: Update when RxExpect support RxSwift 5.0
+    // .package(url: "https://github.com/devxoul/RxExpect.git", .upToNextMajor(from: "1.0.0"))
+    .package(url: "https://github.com/tokijh/RxExpect.git", .branch("master"))
   ],
   targets: [
-    .target(name: "ReactorKit", dependencies: ["ReactorKitRuntime", "RxSwift"]),
+    .target(name: "ReactorKit", dependencies: ["ReactorKitRuntime", "RxSwift", "RxRelay"]),
     .target(name: "ReactorKitRuntime", dependencies: []),
     .testTarget(name: "ReactorKitTests", dependencies: ["ReactorKit", "RxExpect"]),
   ]
